Clean up files, file names, and method names
