+++ src/truncated_input_stream.jl
